Move base class HydeExtension into Concerns namespace
