+++ packages/framework/src/Foundation/PageCollection.php
@@ -88,7 +88,7 @@ final class PageCollection extends BaseFoundationCollection
             $this->discoverPagesFor($pageClass);
         }
 
-        /** @var class-string<\Hyde\Foundation\Extensions\HydeExtension> $extension */
+        /** @var class-string<\Hyde\Foundation\Concerns\HydeExtension> $extension */
         foreach ($this->kernel->getRegisteredExtensions() as $extension) {
             $extension::discoverPages($this);
         }
